graduate_pro/requirements.txt
absl-py==2.3.0
action-msgs==1.2.1
actionlib-msgs==4.8.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
ajsonrpc==1.2.0
altair==5.5.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.14
ament-package==0.14.0
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
bitarray==3.4.2
bottle==0.13.3
Brotli @ file:///croot/brotli-split_1736182456865/work
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.6
coloredlogs==15.0.1
composition-interfaces==1.2.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.49.0
controller-manager-msgs==2.49.0
coverage==7.8.0
cv-bridge==1.13.0.post0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.1.1
decorator==5.2.1
deep-sort-realtime==1.3.2
diagnostic-msgs==4.8.0
digit-interface==0.2.1
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
domain-coordinator==0.10.0
empy==3.3.4
exceptiongroup==1.3.0
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.88.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools @ file:///croot/fonttools_1737039080035/work
frozenlist==1.6.0
fsspec==2025.5.1
generate-parameter-library-py==0.4.0
geometry-msgs==4.8.0
gradio==3.34.0
gradio_client==1.10.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
ikpy==3.4.2
imageio @ file:///croot/imageio_1738159938990/work
iniconfig==2.1.0
interactive-markers==2.3.2
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
joint-state-publisher==2.4.0
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
lark==1.2.2
laser-geometry==2.4.0
launch==1.0.8
launch-param-builder==0.1.1
launch-ros==0.19.9
lazy_loader @ file:///croot/lazy_loader_1718176737906/work
librosa==0.9.1
lifecycle-msgs==1.2.1
linkify-it-py==2.0.3
llvmlite==0.39.0
lxml==5.4.0
map-msgs==2.1.0
Markdown==3.8
markdown-it-py==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
message-filters==4.3.6
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
moveit-configs-utils==2.5.8
moveit-msgs==2.2.1
mpmath==1.3.0
multidict==6.4.4
narwhals==1.41.0
nav-msgs==4.8.0
networkx==3.4.2
notify2==0.3.1
numba==0.56.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
omegaconf==2.3.0
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0
opencv-python-headless==4.10.0
orjson==3.10.18
osrf-pycommon==2.1.6
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
platformio==6.1.18
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
praat-parselmouth==0.4.5
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydub==0.25.1
pyelftools==0.32
Pygments==2.19.1
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6_sip @ file:///croot/pyqt-split_1744804475988/work/pyqt_sip
pyrealsense2==2.55.1.6486
pyserial==3.5
pytest==8.3.5
pytest-cov==6.1.1
pytest-repeat==0.9.4
pytest-rerunfailures==15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyudev==0.24.3
pyworld==0.3.2
PyYAML==6.0.2
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
realsense2-camera-msgs==4.55.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2cli==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
rosdep==0.25.1
rosdistro==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg==1.6.0
rpds-py==0.25.1
rpyutils==0.2.1
sacrebleu==2.5.1
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
scikit-image @ file:///croot/scikit-image_1750419472910/work
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
-e git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf#egg=segment_anything
semantic-version==2.10.0
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1646101165171/work
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
soundfile==0.13.1
srdfdom==2.0.7
starlette==0.22.0
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
threadpoolctl==3.6.0
tifffile==2025.5.10
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.7.0
torchaudio==2.7.0
torchcrepe==0.0.20
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.8.0
transforms3d==0.4.2
triton==3.3.0
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
ultralytics==8.3.119
ultralytics-thop==2.0.14
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.4.0
uvicorn==0.34.3
visualization-msgs==4.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xacro==2.0.13
xarm==0.0.4
xArm-Python-SDK @ file:///home/qi/%E4%B8%8B%E8%BD%BD/xArm-Python-SDK
yarl==1.20.0
